[FIX] All warnings found by using the option BTK_EXTRA_COMPILER_WARNINGS with GCC are cleared. [FIX] In the method btk::Analog::SetGain, the test between the new gain and the current one is wrong.
git-svn-id: https://b-tk.googlecode.com/svn/BTK/trunk@421 93be6e1e-6fcc-11de-baec-69240fdc299a

+++ Code/BasicFilters/btkMergeAcquisitionFilter.cpp
@@ -648,7 +648,7 @@ namespace btk
             itChan = fpChanList.erase(itChan);
             continue;
           }
-          if (channel->GetValues().size() == (chanNumber * usedNumber))
+          if (static_cast<int>(channel->GetValues().size()) == (chanNumber * usedNumber))
           {
             (*itConfig)->AppendChild((*itChan)->GetChild("CHANNEL"));
             this->UpdateForcePlatformMetaData(out, *itConfig);
